SystemUI: show date on 2 lines in status bar
This allows showing of the full day of the week, and removes
the blank space above the date (which always felt strange to
me when clicking it to access the calendar).

It also fixes a spacing issue on smaller screen devices that
causes the quick settings button to be partially chopped off
when the AM/PM clock settings are enabled.

Before: http://imgur.com/UzkFe
After: http://imgur.com/UXoBJ

Signed-off-by: Paul Beeler <sparksco@gmail.com>